Added quadruple bond as option in Surface_Abstraction/Abstracting
If we allow triple bonds in the donating, then we need to 
allow quadruple bonds in the abstracting, so that we can find
the same reaction in both directions (needed to find kinetics).

Unfortunately this doesn't actually work and crashes for 
Carbon quadruple-bonded to the surface.
Stepping through the debugger, either
R is not matching the C, and/or Xo is not matching the surface site.
The reactant.findSubgraphIsomorphisms(struct) is in cythonized code
so I can't inspect inside it, but it seems the surface site has 
atomType 'Xv' and the C has atomType 'Cs'. Probably the atom typing
code doesn't have Q bonds.