Remove forbidden group designed to stop failed bidentate desorption.
This is now handled by hard-coding in RMG-Py commit
5aee27efdd4aef00137d67f8219bccec5fa11843
"Ensure desorption has a gas phase species."
that ensures an adsorption reaction starts with a desorbed species.

Revert "I added a forbidden structure to prevent di-sigma species from breaking one of the bonds. I'm not sure if this is a good idea or not."

This reverts commit 09c0a1994ae15c0dbb3e1be9b1221f146893e9af.