In bidentate adsorbates, the surface sites are not bonded to each other.
We currently can't figure out the atom type for a surface site 
with two bonds (one to another surface site). And the bidentate
adsorption recipe was not joining the sites together anyway,
so any adsorbates made this way would not have the bond.

This commit removes it from the thermo adsorption energy group
estimate definitions (as these are group templates, they should 
match molecules in which the sites *are* bonded to each other,
eg. in reaction libraries, although we should work to remove those).